+++ src/nova/trame/view/components/data_selector.py
@@ -2,14 +2,13 @@
 
 from asyncio import ensure_future, sleep
 from typing import Any, List, Optional, cast
-from warnings import warn
 
 from trame.app import get_server
 from trame.widgets import client, datagrid, html
 from trame.widgets import vuetify3 as vuetify
 
 from nova.mvvm.trame_binding import TrameBinding
-from nova.trame.model.data_selector import CUSTOM_DIRECTORIES_LABEL, DataSelectorModel
+from nova.trame.model.data_selector import DataSelectorModel, DataSelectorState
 from nova.trame.view.layouts import GridLayout, HBoxLayout, VBoxLayout
 from nova.trame.view_model.data_selector import DataSelectorViewModel
 
@@ -19,16 +18,14 @@ vuetify.enable_lab()
 
 
 class DataSelector(datagrid.VGrid):
-    """Allows the user to select datafiles from an IPTS experiment."""
+    """Allows the user to select datafiles from the server."""
 
     REFRESH_RATE = 30
 
     def __init__(
         self,
         v_model: str,
-        allow_custom_directories: bool = False,
-        facility: str = "",
-        instrument: str = "",
+        directory: str,
         extensions: Optional[List[str]] = None,
         prefix: str = "",
         select_strategy: str = "all",
@@ -41,18 +38,14 @@ class DataSelector(datagrid.VGrid):
         v_model : str
             The name of the state variable to bind to this widget. The state variable will contain a list of the files
             selected by the user.
-        allow_custom_directories : bool, optional
-            Whether or not to allow users to provide their own directories to search for datafiles in. Ignored if the
-            facility parameter is set.
-        facility : str, optional
-            The facility to restrict data selection to. Options: HFIR, SNS
-        instrument : str, optional
-            The instrument to restrict data selection to. Please use the instrument acronym (e.g. CG-2).
+        directory : str
+            The top-level folder to expose to users. Only contents of this directory and its children will be exposed to
+            users.
         extensions : List[str], optional
             A list of file extensions to restrict selection to. If unset, then all files will be shown.
         prefix : str, optional
-            A subdirectory within the user's chosen experiment to show files. If not specified, the user will be shown a
-            folder browser and will be able to see all files in the experiment that they have access to.
+            A subdirectory within the selected top-level folder to show files. If not specified, the user will be shown
+            a folder browser and will be able to see all files in the selected top-level folder.
         select_strategy : str, optional
             The selection strategy to pass to the `VDataTable component <https://trame.readthedocs.io/en/latest/trame.widgets.vuetify3.html#trame.widgets.vuetify3.VDataTable>`__.
             If unset, the `all` strategy will be used.
@@ -64,6 +57,12 @@ class DataSelector(datagrid.VGrid):
         -------
         None
         """
+        if "allow_custom_directory" in kwargs or "facility" in kwargs or "instrument" in kwargs:
+            raise TypeError(
+                "The old DataSelector component has been renamed to NeutronDataSelector. Please import it from "
+                "`nova.trame.view.components.ornl`."
+            )
+
         if "items" in kwargs:
             raise AttributeError("The items parameter is not allowed on DataSelector widget.")
 
@@ -72,21 +71,15 @@ class DataSelector(datagrid.VGrid):
         else:
             self._label = None
 
-        if facility and allow_custom_directories:
-            warn("allow_custom_directories will be ignored since the facility parameter is set.", stacklevel=1)
-
         self._v_model = v_model
         self._v_model_name_in_state = v_model.split(".")[0]
-        self._allow_custom_directories = allow_custom_directories
+        self._directory = directory
         self._extensions = extensions if extensions is not None else []
         self._prefix = prefix
         self._select_strategy = select_strategy
 
         self._revogrid_id = f"nova__dataselector_{self._next_id}_rv"
         self._state_name = f"nova__dataselector_{self._next_id}_state"
-        self._facilities_name = f"nova__dataselector_{self._next_id}_facilities"
-        self._instruments_name = f"nova__dataselector_{self._next_id}_instruments"
-        self._experiments_name = f"nova__dataselector_{self._next_id}_experiments"
         self._directories_name = f"nova__dataselector_{self._next_id}_directories"
         self._datafiles_name = f"nova__dataselector_{self._next_id}_datafiles"
 
@@ -96,39 +89,17 @@ class DataSelector(datagrid.VGrid):
         ).exec
         self._reset_state = client.JSEval(exec=f"{self._v_model} = []; {self._flush_state}").exec
 
-        self.create_model(facility, instrument)
+        self.create_model()
         self.create_viewmodel()
 
-        self.create_ui(facility, instrument, **kwargs)
+        self.create_ui(**kwargs)
 
         ensure_future(self._refresh_loop())
 
-    def create_ui(self, facility: str, instrument: str, **kwargs: Any) -> None:
-        with VBoxLayout(classes="nova-data-selector", height="100%"):
+    def create_ui(self, *args: Any, **kwargs: Any) -> None:
+        with VBoxLayout(classes="nova-data-selector", height="100%") as self._layout:
             with HBoxLayout(valign="center"):
-                with GridLayout(classes="flex-1-1", columns=3):
-                    columns = 3
-                    if facility == "":
-                        columns -= 1
-                        InputField(
-                            v_model=f"{self._state_name}.facility", items=(self._facilities_name,), type="autocomplete"
-                        )
-                    if instrument == "":
-                        columns -= 1
-                        InputField(
-                            v_if=f"{self._state_name}.facility !== '{CUSTOM_DIRECTORIES_LABEL}'",
-                            v_model=f"{self._state_name}.instrument",
-                            items=(self._instruments_name,),
-                            type="autocomplete",
-                        )
-                    InputField(
-                        v_if=f"{self._state_name}.facility !== '{CUSTOM_DIRECTORIES_LABEL}'",
-                        v_model=f"{self._state_name}.experiment",
-                        column_span=columns,
-                        items=(self._experiments_name,),
-                        type="autocomplete",
-                    )
-                    InputField(v_else=True, v_model=f"{self._state_name}.custom_directory", column_span=2)
+                self._layout.filter = html.Div(classes="flex-1-1")
                 with vuetify.VBtn(
                     classes="mx-1", density="compact", icon=True, variant="text", click=self.refresh_file_list
                 ):
@@ -148,7 +119,7 @@ class DataSelector(datagrid.VGrid):
                             item_value="path",
                             items=(self._directories_name,),
                             click_open=(self._vm.expand_directory, "[$event.path]"),
-                            update_activated=(self._vm.set_directory, "$event"),
+                            update_activated=(self._vm.set_subdirectory, "$event"),
                         )
                         vuetify.VListItem("No directories found", classes="flex-0-1 text-center", v_else=True)
 
@@ -209,10 +180,9 @@ class DataSelector(datagrid.VGrid):
                         f"(+{{{{ {self._v_model}.length - 2 }}}} others)", v_if="index === 2", classes="text-caption"
                     )
 
-    def create_model(self, facility: str, instrument: str) -> None:
-        self._model = DataSelectorModel(
-            facility, instrument, self._extensions, self._prefix, self._allow_custom_directories
-        )
+    def create_model(self) -> None:
+        state = DataSelectorState()
+        self._model = DataSelectorModel(state, self._directory, self._extensions, self._prefix)
 
     def create_viewmodel(self) -> None:
         server = get_server(None, client_type="vue3")
@@ -220,9 +190,6 @@ class DataSelector(datagrid.VGrid):
 
         self._vm = DataSelectorViewModel(self._model, binding)
         self._vm.state_bind.connect(self._state_name)
-        self._vm.facilities_bind.connect(self._facilities_name)
-        self._vm.instruments_bind.connect(self._instruments_name)
-        self._vm.experiments_bind.connect(self._experiments_name)
         self._vm.directories_bind.connect(self._directories_name)
         self._vm.datafiles_bind.connect(self._datafiles_name)
         self._vm.reset_bind.connect(self.reset)
@@ -236,29 +203,11 @@ class DataSelector(datagrid.VGrid):
         self._reset_state()
         self._reset_rv_grid()
 
-    def set_state(
-        self, facility: Optional[str] = None, instrument: Optional[str] = None, experiment: Optional[str] = None
-    ) -> None:
-        """Programmatically set the facility, instrument, and/or experiment to restrict data selection to.
-
-        If a parameter is None, then it will not be updated.
-
-        Parameters
-        ----------
-        facility : str, optional
-            The facility to restrict data selection to. Options: HFIR, SNS
-        instrument : str, optional
-            The instrument to restrict data selection to. Must be at the selected facility.
-        experiment : str, optional
-            The experiment to restrict data selection to. Must begin with "IPTS-". It is your responsibility to validate
-            that the provided experiment exists within the instrument directory. If it doesn't then no datafiles will be
-            shown to the user.
-
-        Returns
-        -------
-        None
-        """
-        self._vm.set_state(facility, instrument, experiment)
+    def set_state(self, *args: Any, **kwargs: Any) -> None:
+        raise TypeError(
+            "The old DataSelector component has been renamed to NeutronDataSelector. Please import it from "
+            "`nova.trame.view.components.ornl`."
+        )
 
     async def _refresh_loop(self) -> None:
         while True: