+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+### nova-trame, 0.23.0
+
+* The existing DataSelector component has been renamed to NeutronDataSelector and moved to nova.trame.view.components.ornl.NeutronDataSelector (thanks to John Duggan).
+
 ### nova-trame, 0.22.1
 
-* Themed app now has an Exit Button by default which closes the application and can stop any running jobs (thanks to Gregory Cage).
+* ThemedApp now has an Exit Button by default which closes the application and can stop any running jobs (thanks to Gregory Cage).
 
 ### nova-trame, 0.22.0